+++ client/src/Utils/types.ts
@@ -15,6 +15,18 @@ export const OH_CONFIG_PARAMETERS = {
         remoteEnabled: 'languageserver.remoteEnabled',
         remotePort: 'languageserver.remotePort',
     },
+    log: {
+        eventsLogPath: 'log.eventsLogPath',
+        openhabLogPath: 'log.openhabLogPath',
+    },
+    hover: {
+        showItemDefinition: 'hover.showItemDefinition',
+        showRuleReferences: 'hover.showRuleReferences',
+        showSitemapReferences: 'hover.showSitemapReferences',
+        showScriptReferences: 'hover.showScriptReferences',
+        showLogSearch: 'hover.showLogSearch',
+        maxReferencesPerCategory: 'hover.maxReferencesPerCategory',
+    },
     itemCasing: 'itemCasing',
     consoleCommand: 'consoleCommand',
     useRestApi: 'useRestApi',